+++ naga/src/valid/immediates.rs
@@ -1,11 +1,69 @@
-use core::{fmt, ops};
+use core::{
+    fmt::{self, Debug},
+    ops,
+};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, thiserror::Error)]
+#[error("Immediate size {0} overflows the bitmask")]
+pub struct ImmediateSlotsOverflowError(pub u64);
 
 /// A bitmask, tracking which 4-byte slots have been written via `set_immediates`.
 /// Bit N corresponds to bytes [N*4 .. N*4+4).
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize))]
-#[cfg_attr(feature = "deserialize", derive(serde::Deserialize))]
+#[derive(Clone, Copy, Default, PartialEq, Eq)]
 pub struct ImmediateSlots(u64);
+
+#[derive(Clone, Copy, Debug)]
+pub enum ImmediateUsage {
+    Valid { slots: ImmediateSlots, size: u32 },
+    // Used when a shader's immediate data type exceeds the maximum of 256 bytes and cannot
+    // be represented in the `ImmediateSlots` bitmask.
+    Invalid { size: u32 },
+}
+
+impl Default for ImmediateUsage {
+    fn default() -> Self {
+        ImmediateUsage::Valid {
+            slots: ImmediateSlots::default(),
+            size: 0,
+        }
+    }
+}
+
+impl ImmediateUsage {
+    pub const fn size(&self) -> u32 {
+        match *self {
+            ImmediateUsage::Valid { size, .. } => size,
+            ImmediateUsage::Invalid { size } => size,
+        }
+    }
+
+    pub fn from_type(
+        ty: &crate::TypeInner,
+        types: &crate::UniqueArena<crate::Type>,
+        gctx: crate::proc::GlobalCtx,
+    ) -> Self {
+        let size = ty.size(gctx);
+        ImmediateSlots::from_type(ty, types, gctx)
+            .map(|slots| Self::Valid { slots, size })
+            .unwrap_or(Self::Invalid { size })
+    }
+
+    pub fn merge(&self, other: &ImmediateUsage) -> Self {
+        let size = self.size().max(other.size());
+        match (*self, *other) {
+            (
+                ImmediateUsage::Valid { slots, .. },
+                ImmediateUsage::Valid {
+                    slots: other_slots, ..
+                },
+            ) => Self::Valid {
+                slots: slots | other_slots,
+                size,
+            },
+            _ => Self::Invalid { size },
+        }
+    }
+}
 
 impl ImmediateSlots {
     pub const fn from_raw(raw: u64) -> Self {
@@ -13,47 +71,69 @@ impl ImmediateSlots {
     }
 
     /// Compute the bitmask for a byte range [offset .. offset + size_bytes).
-    pub const fn from_range(offset: u32, size_bytes: u32) -> Self {
+    pub const fn from_range(
+        offset: u32,
+        size_bytes: u32,
+    ) -> Result<Self, ImmediateSlotsOverflowError> {
+        let Some(end) = offset.checked_add(size_bytes) else {
+            return Err(ImmediateSlotsOverflowError(
+                offset as u64 + size_bytes as u64,
+            ));
+        };
+        if end > u64::BITS * 4 {
+            return Err(ImmediateSlotsOverflowError(end as u64));
+        }
         if size_bytes == 0 {
-            return Self(0);
+            return Ok(Self(0));
         }
         let lo = offset / 4;
         let hi = (offset + size_bytes).div_ceil(4);
-        Self(u64::MAX << lo & u64::MAX >> (64 - hi))
+        Ok(Self(u64::MAX << lo & u64::MAX >> (64 - hi)))
     }
 
-    /// Compute the slots occupied by a type at a given byte offset,
-    /// excluding padding between struct members.
+    /// Compute the slots occupied by a type,
+    /// excluding padding between matrix columns or struct members.
     pub fn from_type(
         ty: &crate::TypeInner,
-        offset: u32,
         types: &crate::UniqueArena<crate::Type>,
         gctx: crate::proc::GlobalCtx,
-    ) -> Self {
-        // <https://www.w3.org/TR/WGSL/#accessible-bytes>
-        match *ty {
-            crate::TypeInner::Matrix {
-                columns,
-                rows,
-                scalar,
-            } => {
-                let mut slots = Self::default();
-                let stride = crate::proc::Alignment::from(rows) * scalar.width as u32;
-                for col in 0..u32::from(columns) {
-                    slots |= Self::from_range(col * stride, u32::from(rows) * scalar.width as u32);
+    ) -> Result<Self, ImmediateSlotsOverflowError> {
+        fn from_type_recursive(
+            ty: &crate::TypeInner,
+            offset: u32,
+            types: &crate::UniqueArena<crate::Type>,
+            gctx: crate::proc::GlobalCtx,
+        ) -> Result<ImmediateSlots, ImmediateSlotsOverflowError> {
+            // <https://www.w3.org/TR/WGSL/#accessible-bytes>
+            match *ty {
+                crate::TypeInner::Matrix {
+                    columns,
+                    rows,
+                    scalar,
+                } => {
+                    let mut slots = ImmediateSlots::default();
+                    let stride = crate::proc::Alignment::from(rows) * u32::from(scalar.width);
+                    for col in 0..u32::from(columns) {
+                        slots |= ImmediateSlots::from_range(
+                            offset + col * stride,
+                            u32::from(rows) * u32::from(scalar.width),
+                        )?;
+                    }
+                    Ok(slots)
                 }
-                slots
-            }
-            crate::TypeInner::Struct { ref members, .. } => {
-                let mut slots = Self::default();
-                for member in members {
-                    let member_ty = &types[member.ty].inner;
-                    slots |= Self::from_type(member_ty, offset + member.offset, types, gctx);
+                crate::TypeInner::Struct { ref members, .. } => {
+                    let mut slots = ImmediateSlots::default();
+                    for member in members {
+                        let member_ty = &types[member.ty].inner;
+                        slots |=
+                            from_type_recursive(member_ty, offset + member.offset, types, gctx)?;
+                    }
+                    Ok(slots)
                 }
-                slots
+                _ => ImmediateSlots::from_range(offset, ty.size(gctx)),
             }
-            _ => Self::from_range(offset, ty.size(gctx)),
         }
+        from_type_recursive(ty, 0, types, gctx)
     }
 
     /// Returns true if `self` contains all bits in `other`.
@@ -64,46 +144,6 @@ impl ImmediateSlots {
     /// Returns the bits in `self` that are not set in `other`.
     pub const fn difference(self, other: Self) -> Self {
         Self(self.0 & !other.0)
-    }
-
-    /// Compute the immediate slot bitmask for a pointer expression that
-    /// refers to (part of) an immediate global variable.
-    ///
-    /// `global` is the handle of the immediate global variable that this
-    /// pointer derives from (obtained from `assignable_global`).
-    pub(crate) fn for_pointer(
-        pointer: crate::arena::Handle<crate::Expression>,
-        global: crate::arena::Handle<crate::GlobalVariable>,
-        expression_arena: &crate::Arena<crate::Expression>,
-        global_vars: &crate::Arena<crate::GlobalVariable>,
-        types: &crate::UniqueArena<crate::Type>,
-    ) -> Self {
-        use crate::Expression as E;
-        use crate::TypeInner;
-
-        let gctx = crate::proc::GlobalCtx {
-            types,
-            constants: &crate::Arena::new(),
-            overrides: &crate::Arena::new(),
-            global_expressions: &crate::Arena::new(),
-        };
-
-        let global_ty = &types[global_vars[global].ty].inner;
-
-        match expression_arena[pointer] {
-            E::GlobalVariable(_) => Self::from_type(global_ty, 0, types, gctx),
-            E::AccessIndex { base, index } => {
-                if let E::GlobalVariable(_) = expression_arena[base] {
-                    if let TypeInner::Struct { ref members, .. } = *global_ty {
-                        let member = &members[index as usize];
-                        let member_ty = &types[member.ty].inner;
-                        return Self::from_type(member_ty, member.offset, types, gctx);
-                    }
-                }
-                Self::from_type(global_ty, 0, types, gctx)
-            }
-            _ => Self::from_type(global_ty, 0, types, gctx),
-        }
     }
 }
 
@@ -147,22 +187,30 @@ impl fmt::Display for ImmediateSlots {
     }
 }
 
+impl Debug for ImmediateSlots {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{self}")
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use crate::valid::ImmediateSlotsOverflowError;
+
     use super::ImmediateSlots;
 
     #[test]
     fn range_single() {
         assert_eq!(
-            ImmediateSlots::from_range(0, 4),
+            ImmediateSlots::from_range(0, 4).unwrap(),
             ImmediateSlots::from_raw(0b1)
         );
         assert_eq!(
-            ImmediateSlots::from_range(4, 4),
+            ImmediateSlots::from_range(4, 4).unwrap(),
             ImmediateSlots::from_raw(0b10)
         );
         assert_eq!(
-            ImmediateSlots::from_range(8, 4),
+            ImmediateSlots::from_range(8, 4).unwrap(),
             ImmediateSlots::from_raw(0b100)
         );
     }
@@ -170,11 +218,11 @@ mod tests {
     #[test]
     fn range_vec4() {
         assert_eq!(
-            ImmediateSlots::from_range(0, 16),
+            ImmediateSlots::from_range(0, 16).unwrap(),
             ImmediateSlots::from_raw(0b1111)
         );
         assert_eq!(
-            ImmediateSlots::from_range(16, 16),
+            ImmediateSlots::from_range(16, 16).unwrap(),
             ImmediateSlots::from_raw(0b1111_0000)
         );
     }
@@ -182,9 +230,36 @@ mod tests {
     #[test]
     fn range_full_256() {
         assert_eq!(
-            ImmediateSlots::from_range(0, 256),
+            ImmediateSlots::from_range(0, 256).unwrap(),
             ImmediateSlots::from_raw(u64::MAX)
         );
+    }
+
+    #[test]
+    fn range_overflow() {
+        assert_eq!(
+            ImmediateSlots::from_range(0, 257),
+            Err(ImmediateSlotsOverflowError(257))
+        );
+    }
+
+    #[test]
+    fn from_type_overflow() {
+        let module = crate::front::wgsl::parse_str(
+            "struct S { \
+            e64: mat4x4<f32>, \
+            e128: mat4x4<f32>, \
+            e192: mat4x4<f32>, \
+            e256: mat4x4<f32>, \
+            e260: f32\
+            }",
+        )
+        .unwrap();
+        let struct_ty = (module.types.iter().map(|ty| ty.1))
+            .find(|ty| ty.name.as_deref() == Some("S"))
+            .unwrap();
+        let slots = ImmediateSlots::from_type(&struct_ty.inner, &module.types, module.to_ctx());
+        assert_eq!(slots, Err(ImmediateSlotsOverflowError(260)));
     }
 
     #[test]
@@ -193,7 +268,8 @@ mod tests {
         let struct_ty = (module.types.iter().map(|ty| ty.1))
             .find(|ty| ty.name.as_deref() == Some("S"))
             .unwrap();
-        let slots = ImmediateSlots::from_type(&struct_ty.inner, 0, &module.types, module.to_ctx());
+        let slots =
+            ImmediateSlots::from_type(&struct_ty.inner, &module.types, module.to_ctx()).unwrap();
         assert_eq!(slots, ImmediateSlots::from_raw(0b1111_0001));
     }
 
@@ -203,18 +279,28 @@ mod tests {
         let struct_ty = (module.types.iter().map(|ty| ty.1))
             .find(|ty| ty.name.as_deref() == Some("S"))
             .unwrap();
-        let slots = ImmediateSlots::from_type(&struct_ty.inner, 0, &module.types, module.to_ctx());
+        let slots =
+            ImmediateSlots::from_type(&struct_ty.inner, &module.types, module.to_ctx()).unwrap();
         assert_eq!(slots, ImmediateSlots::from_raw(0b0111_0111_0111));
+
+        let module =
+            crate::front::wgsl::parse_str("struct S { f: f32, mat: mat2x2<f32> }").unwrap();
+        let struct_ty = (module.types.iter().map(|ty| ty.1))
+            .find(|ty| ty.name.as_deref() == Some("S"))
+            .unwrap();
+        let slots =
+            ImmediateSlots::from_type(&struct_ty.inner, &module.types, module.to_ctx()).unwrap();
+        assert_eq!(slots, ImmediateSlots::from_raw(0b11_11_01));
     }
 
     #[test]
     fn range_unaligned() {
         assert_eq!(
-            ImmediateSlots::from_range(0, 3),
+            ImmediateSlots::from_range(0, 3).unwrap(),
             ImmediateSlots::from_raw(0b1)
         );
         assert_eq!(
-            ImmediateSlots::from_range(0, 5),
+            ImmediateSlots::from_range(0, 5).unwrap(),
             ImmediateSlots::from_raw(0b11)
         );
     }
@@ -224,16 +310,16 @@ mod tests {
         let required = ImmediateSlots::from_raw(0b1111_0001);
         let mut set = ImmediateSlots::default();
         assert!(!set.contains(required));
-        set |= ImmediateSlots::from_range(0, 4);
+        set |= ImmediateSlots::from_range(0, 4).unwrap();
         assert!(!set.contains(required));
-        set |= ImmediateSlots::from_range(16, 16);
+        set |= ImmediateSlots::from_range(16, 16).unwrap();
         assert!(set.contains(required));
     }
 
     #[test]
     fn difference() {
         let required = ImmediateSlots::from_raw(0b1111_0001);
-        let set = ImmediateSlots::from_range(0, 4);
+        let set = ImmediateSlots::from_range(0, 4).unwrap();
         assert_eq!(
             required.difference(set),
             ImmediateSlots::from_raw(0b1111_0000)