+++ tests/tests/wgpu-gpu/resource_error.rs
@@ -1,11 +1,11 @@
 use wgpu_test::{fail, gpu_test, valid, GpuTestConfiguration, GpuTestInitializer, TestParameters};
 
 pub fn all_tests(vec: &mut Vec<GpuTestInitializer>) {
-    vec.extend([BAD_BUFFER, BAD_TEXTURE]);
+    vec.extend([BAD_BUFFER_MAP, BAD_TEXTURE_CREATE_VIEW, BAD_TEXTURE_WRITE]);
 }
 
 #[gpu_test]
-static BAD_BUFFER: GpuTestConfiguration = GpuTestConfiguration::new()
+static BAD_BUFFER_MAP: GpuTestConfiguration = GpuTestConfiguration::new()
     .parameters(TestParameters::default().enable_noop())
     .run_sync(|ctx| {
         // Create a buffer with bad parameters and call a few methods.
@@ -38,7 +38,7 @@ static BAD_BUFFER: GpuTestConfiguration = GpuTestConfiguration::new()
     });
 
 #[gpu_test]
-static BAD_TEXTURE: GpuTestConfiguration = GpuTestConfiguration::new()
+static BAD_TEXTURE_CREATE_VIEW: GpuTestConfiguration = GpuTestConfiguration::new()
     .parameters(TestParameters::default().enable_noop())
     .run_sync(|ctx| {
         let texture = fail(
@@ -71,4 +71,52 @@ static BAD_TEXTURE: GpuTestConfiguration = GpuTestConfiguration::new()
         );
         valid(&ctx.device, || texture.destroy());
         valid(&ctx.device, || texture.destroy());
+    });
+
+#[gpu_test]
+static BAD_TEXTURE_WRITE: GpuTestConfiguration = GpuTestConfiguration::new()
+    .parameters(TestParameters::default().enable_noop())
+    .run_sync(|ctx| {
+        let texture = fail(
+            &ctx.device,
+            || {
+                ctx.device.create_texture(&wgpu::TextureDescriptor {
+                    label: None,
+                    size: wgpu::Extent3d {
+                        width: 256,
+                        height: 256,
+                        depth_or_array_layers: 1,
+                    },
+                    mip_level_count: 5888,
+                    sample_count: 1,
+                    dimension: wgpu::TextureDimension::D2,
+                    format: wgpu::TextureFormat::Rgba8Unorm,
+                    usage: wgpu::TextureUsages::COPY_DST,
+                    view_formats: &[],
+                })
+            },
+            Some("mip level count"),
+        );
+
+        fail(
+            &ctx.device,
+            || {
+                ctx.queue.write_texture(
+                    wgpu::TexelCopyTextureInfo {
+                        texture: &texture,
+                        mip_level: 255,
+                        origin: wgpu::Origin3d::ZERO,
+                        aspect: wgpu::TextureAspect::All,
+                    },
+                    &[0u8; 4],
+                    wgpu::TexelCopyBufferLayout::default(),
+                    wgpu::Extent3d {
+                        width: 1,
+                        height: 1,
+                        depth_or_array_layers: 1,
+                    },
+                );
+            },
+            Some("Texture with '' label is invalid"),
+        );
     });